requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Cerberus==1.3.4
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
